requirements.txt
requests==2.22.0
django_annoying==0.10.5
Django==2.2.5
Markdown==3.1.1
djangorestframework==3.10.3
django-simpleui==2.9
django-ckeditor==5.7.1
drf-extensions==0.5.0